move TrackerEventDiffEngine to `trackers` module
